+++ src/hx711.c
@@ -79,7 +79,7 @@ void hx711_init(
 
         //either statement below will panic if it fails
         hx->_offset = pio_add_program(hx->_pio, hx->_prog);
-        hx->_state_mach = pio_claim_unused_sm(hx->_pio, true);
+        hx->_state_mach = (uint)pio_claim_unused_sm(hx->_pio, true);
 
         prog_init_func(hx);
